refactor: consolidate api imports and exports for consistency across modules
