+++ src/popup/components/ai-move/use-ai-move.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { useMemoizedFn } from 'ahooks'
 import { useShallow } from 'zustand/react/shallow'
-import { fetchAllFavoriteMedias, queryAndSendMessage } from '@/utils/tab'
+import { queryAndSendMessage } from '@/utils/tab'
+import { fetchAllFavoriteMedias } from '@/utils/api'
 import { MessageEnum } from '@/utils/message'
 import { createStreamAdapter } from '@/hooks/use-create-keyword-by-ai/ai-stream-parser'
 import { useGlobalConfig } from '@/store/global-data'
